+++ src/managers/SingleManager.js
@@ -8,6 +8,7 @@ import { terminal } from "@absolunet/terminal";
 import builder from "../helpers/builder.js";
 import documenter from "../helpers/documenter.js";
 import paths from "../helpers/paths.js";
+import reformatter from "../helpers/reformatter.js";
 import util from "../helpers/util.js";
 import AbstractManager from "./AbstractManager.js";
 
@@ -64,6 +65,16 @@ class SingleManager extends AbstractManager {
 		return super.watch(options, async () => {
 			// Run watcher
 			await builder.watch(__(this).get("dist"));
+		});
+	}
+
+	/**
+	 * @inheritdoc
+	 */
+	reformat(options) {
+		return super.reformat(options, async () => {
+			// Run reformatter
+			await reformatter.run();
 		});
 	}