+++ src/Manager.js
@@ -159,8 +159,8 @@ class Manager {
 			case environment.TASK.watch:
 				await managerType.watch();
 				break;
-			case environment.TASK.reformat:
-				await managerType.reformat();
+			case environment.TASK.fix:
+				await managerType.fix();
 				break;
 			case environment.TASK.documentation:
 				await managerType.documentation();