+++ src/managers/AbstractManager.js
@@ -159,7 +159,7 @@ class AbstractManager {
 	}
 
 	/**
-	 * Reformat task.
+	 * Fix task.
 	 *
 	 * @async
 	 * @param {object} [options] - Options.
@@ -167,14 +167,14 @@ class AbstractManager {
 	 * @param {Function} [toExecute] - Async function to execute.
 	 * @returns {Promise} When task completed.
 	 */
-	reformat(
+	fix(
 		{ grouped } = {},
 		toExecute = async () => {
 			/**/
 		}
 	) {
 		return runTask({
-			task: "reformat",
+			task: "fix",
 			context: this,
 			grouped,
 			toExecute,
@@ -250,7 +250,7 @@ class AbstractManager {
 			context: this,
 			grouped,
 			toExecute: async () => {
-				await this.reformat({ grouped: true });
+				await this.fix({ grouped: true });
 				await this.build({ grouped: true });
 				await this.prepare({ grouped: true });
 				await this.documentation({ grouped: true });