+++ src/managers/MultiManager.js
@@ -8,8 +8,8 @@ import __ from "@absolunet/private-registry";
 import { terminal } from "@absolunet/terminal";
 import builder from "../helpers/builder.js";
 import documenter from "../helpers/documenter.js";
+import fixer from "../helpers/fixer.js";
 import paths from "../helpers/paths.js";
-import reformatter from "../helpers/reformatter.js";
 import util from "../helpers/util.js";
 import AbstractManager from "./AbstractManager.js";
 
@@ -145,10 +145,10 @@ class MultiManager extends AbstractManager {
 	/**
 	 * @inheritdoc
 	 */
-	reformat(options) {
-		return super.reformat(options, async () => {
-			// Run reformatter
-			await reformatter.run();
+	fix(options) {
+		return super.fix(options, async () => {
+			// Run fixer
+			await fixer.run();
 		});
 	}