+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add `reformat` task which runs Prettier on all package files
 - In Node.js distribution, use `package.json` engine to configure Babel build
 
 ### Changed